refactor(nimbis): split CLI args into dedicated module (#163)
